% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVQ.R
\name{getCVQ}
\alias{getCVQ}
\title{Calling and Vocational Questionnaire Function}
\usage{
getCVQ(
  RawDataSet,
  longformat = FALSE,
  condensed = TRUE,
  cutoff = 0.8,
  reportmissing = TRUE,
  reverse = TRUE,
  new_id = FALSE,
  drop_id = FALSE
)
}
\arguments{
\item{RawDataSet}{Required DataFrame or Matrix. The R object containing desired items.}

\item{longformat}{Optional Boolean (defaults FALSE) indicating whether to return the CVQ items in a wide of long format.}

\item{condensed}{Optional Boolean (defaults TRUE). Indicates whether the scale and item should be combined (e.g., "CVQ18") or in seperate columns (e.g., "CVQ" and "18"). Note that this ONLY applies when longformat = TRUE.}

\item{cutoff}{Optional Numerical, 0 - 1 (defaults 0.8) Indicates the percentage of required responses be a "qualified" case. For instance, if cutoff = 0.8 then they must have answered at >=80% of the possible items. Enter "0" to include all entries.}

\item{reportmissing}{Optional Boolean (defaults TRUE). Indicates whether to include "nMissing" and "pMissing" columns to indicate the number and percentage of items missing, respectively.}

\item{reverse}{Optional Boolean (defaults TRUE). Indicates whether to reverse score the CVQ-8 item or not.}

\item{new_id}{Optional Boolean (defaults FALSE). Indicates whether to combine year and ID into a new ID formatted as Year_ID.}

\item{drop_id}{Optional Boolean (defaults FALSE). Indicates whether to drop the original ID column.Note that this ONLY applies when new_id = TRUE.}
}
\description{
This function returns a dataframe with the CVQ scores for each participant. Users can
}
\examples{
getCVQ()
getCVQ()
}
\keyword{CVQ}
