% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVQ.R
\name{getCVQ}
\alias{getCVQ}
\title{Calling and Vocational Questionnaire Function}
\usage{
getCVQ(
  RawDataSet,
  longformat = FALSE,
  condensed = TRUE,
  cutoff = 0.8,
  reportmissing = TRUE,
  new_id = FALSE,
  drop_id = FALSE,
  reverse = TRUE,
  subscales = 0
)
}
\arguments{
\item{RawDataSet}{Required DataFrame or Matrix. The R object containing desired items.}

\item{longformat}{Optional Boolean (defaults FALSE) indicating whether to return the CVQ items in a wide of long format.}

\item{condensed}{Optional Boolean (defaults TRUE). Indicates whether the scale and item should be combined (e.g., "CVQ18") or in seperate columns (e.g., "CVQ" and "18"). Note that this ONLY applies when longformat = TRUE.}

\item{cutoff}{Optional Numerical, 0 - 1 (defaults 0.8) Indicates the percentage of required responses be a "qualified" case. For instance, if cutoff = 0.8 then they must have answered at >=80 percent of the possible items. Enter "0" to include all entries.}

\item{reportmissing}{Optional Boolean (defaults TRUE). Indicates whether to include "nMissing" and "pMissing" columns to indicate the number and percentage of items missing, respectively.}

\item{new_id}{Optional Boolean (defaults FALSE). Indicates whether to combine year and ID into a new ID formatted as Year_ID.}

\item{drop_id}{Optional Boolean (defaults FALSE). Indicates whether to drop the original ID column.Note that this ONLY applies when new_id = TRUE.}

\item{reverse}{Optional Boolean (defaults TRUE). Indicates whether to reverse score the CVQ-8 item or not.}

\item{subscales}{Optional numeric (defaults '0'). Indicates the subscales to calculate. Valid options are 0, 2, 6, or 8. The 8-factor option simply returns both the 2- and 6-factor scales. The 0 option doesn't return any subscales. Be aware that item 8 MUST be reverse scored to calculate correctly; you may use "Reverse = TRUE" to do it here, or leave "Reverse = FALSE" if you did it before importing the data. See Dik et al (2012) for details on scoring.}
}
\description{
This function returns a dataframe with the CVQ scores for each participant. Users can
}
\examples{
getCVQ()
}
\keyword{CVQ}
